requirements.txt
asgiref==3.8.1
azure-core==1.34.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dj-database-url==3.0.0
Django==5.2.3
django-storages==1.14.6
django-widget-tweaks==1.5.0
httplib2==0.22.0
idna==3.10
isodate==0.7.2
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
